packet sender divided
