+++ src/pkt_kit/mod.rs
@@ -7,8 +7,11 @@ pub use pkt_builder::PacketBuilder;
 pub mod pkt_dissector;
 pub use pkt_dissector::PacketDissector;
 
-pub mod pkt_sender;
-pub use pkt_sender::PacketSender;
+pub mod layer2_pkt_sender;
+pub use layer2_pkt_sender::Layer2PacketSender;
+
+pub mod layer3_pkt_sender;
+pub use layer3_pkt_sender::Layer3PacketSender;
 
 pub mod pkt_sniffer;
 pub use pkt_sniffer::PacketSniffer;
